% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{acquisitionTime,peakPantheRAnnotation-method}
\alias{acquisitionTime,peakPantheRAnnotation-method}
\alias{acquisitionTime}
\title{acquisitionTime accessor returns value as.POSIXct}
\usage{
\S4method{acquisitionTime}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
acquisitionTime accessor returns value as.POSIXct
}
