% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{annotation_fit_summary_UI_helper}
\alias{annotation_fit_summary_UI_helper}
\title{UI diagnostic table - fit summary}
\usage{
annotation_fit_summary_UI_helper(annot)
}
\arguments{
\item{annot}{(peakPantheRAnnotation) Annotation object}
}
\value{
(data.frame) Fit statistics
}
\description{
Return a table of fit statistic (ratio of peaks found, ratio of peaks filled,
ppm error, RT deviation)
}
