% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{annotation_showMethod_UI_helper}
\alias{annotation_showMethod_UI_helper}
\title{UI show annotation helper - list of properties}
\usage{
annotation_showMethod_UI_helper(annotation)
}
\arguments{
\item{annotation}{(peakPantherAnnotation) Object to describe}
}
\value{
(list) Named list of annotation properties
}
\description{
show method specific to the UI, that returns each field in a named list to
ease display
}
