% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnnotationDiagnosticMultiplot.R
\name{annotationDiagnosticMultiplot}
\alias{annotationDiagnosticMultiplot}
\title{Generate a multiplot of all diagnostic plots}
\usage{
annotationDiagnosticMultiplot(annotationDiagnosticPlotList)
}
\arguments{
\item{annotationDiagnosticPlotList}{(list) List of (one per compound) of list
of diagnostic plots as generated by \code{annotationDiagnosticPlots()}}
}
\value{
A list of multiplots (one per compound)
}
\description{
Generate a multiplot of all diagnostic plots (as generated by
\code{annotationDiagnosticPlots()}) for each compound
}
