% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{annotationParamsDiagnostic,peakPantheRAnnotation-method}
\alias{annotationParamsDiagnostic,peakPantheRAnnotation-method}
\alias{annotationParamsDiagnostic}
\title{Set uROI and FIR based on annotation results
Set updated ROI (uROI) and Fallback Integration Regions (FIR) based on the annotation results. If the object is not annotated, it is returned untouched. ROI is not modified. If uROI exist they are left untouched, otherwise they are set as the minimum and maximum found peaks limits (+/-5\% of ROI in retention time). If FIR are used they are left untouched, otherwise they are set as the median of the found limits (rtMin, rtMax, mzMin, mzMax).}
\usage{
\S4method{annotationParamsDiagnostic}{peakPantheRAnnotation}(object,
  verbose)
}
\arguments{
\item{object}{(peakPantheRAnnotation) Annotated peakPantheRAnnotation object}

\item{verbose}{(bool) If TRUE message progress of uROI and FIR calculation}
}
\value{
(peakPantheRAnnotation) object with updated ROI and FIR set from annotation results
}
\description{
Set uROI and FIR based on annotation results
Set updated ROI (uROI) and Fallback Integration Regions (FIR) based on the annotation results. If the object is not annotated, it is returned untouched. ROI is not modified. If uROI exist they are left untouched, otherwise they are set as the minimum and maximum found peaks limits (+/-5\% of ROI in retention time). If FIR are used they are left untouched, otherwise they are set as the median of the found limits (rtMin, rtMax, mzMin, mzMax).
}
