% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{annotationTable,peakPantheRAnnotation-method}
\alias{annotationTable,peakPantheRAnnotation-method}
\alias{annotationTable}
\title{annotationTable accessor
annotationTable returns a dataframe (row samples, col compounds) filled with a specific peakTable column}
\usage{
\S4method{annotationTable}{peakPantheRAnnotation}(object, column)
}
\arguments{
\item{object}{peakPantheRAnnotation}

\item{column}{a peakTable columns}
}
\value{
(data.frame) (row samples, col compounds) filled with a specific peakTable column
}
\description{
annotationTable accessor
annotationTable returns a dataframe (row samples, col compounds) filled with a specific peakTable column
}
