% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{cpdMetadata,peakPantheRAnnotation-method}
\alias{cpdMetadata,peakPantheRAnnotation-method}
\alias{cpdMetadata}
\title{cpdMetadata accessor}
\usage{
\S4method{cpdMetadata}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
cpdMetadata accessor
}
