% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{cpdName,peakPantheRAnnotation-method}
\alias{cpdName,peakPantheRAnnotation-method}
\alias{cpdName}
\title{cpdName accessor}
\usage{
\S4method{cpdName}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
cpdName accessor
}
