% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{EICs,peakPantheRAnnotation-method}
\alias{EICs,peakPantheRAnnotation-method}
\alias{EICs}
\title{EICs accessor}
\usage{
\S4method{EICs}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
EICs accessor
}
