% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSignalRawData.R
\name{extractSignalRawData}
\alias{extractSignalRawData}
\title{Extract signal in a defined mz rt window from a raw data file}
\usage{
extractSignalRawData(rawSpec, rt, mz, msLevel = 1L)
}
\arguments{
\item{rawSpec}{an \code{\link[MSnbase]{OnDiskMSnExp-class}}}

\item{rt}{(numeric(2)) the retention time range from which the data should be extracted}

\item{mz}{(numeric(2)) the mass range from which the data should be extracted}

\item{msLevel}{(int) the MS level at which the data should be extracted (default to MS level 1)}
}
\value{
a data.frame with signal as row and retention time ("rt"), mass ("mz") and intensity ("int) as columns
}
\description{
Extract all signal from a defined mz rt window from raw data and returns a data.frame. If no rt-mz window is provided, all signal in the raw data file are returned
}
