% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{filename,peakPantheRAnnotation-method}
\alias{filename,peakPantheRAnnotation-method}
\alias{filename}
\title{filename accessor by spliting filepath}
\usage{
\S4method{filename}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(str) filename
}
\description{
filename accessor by spliting filepath
}
