% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFit_utils.R
\name{fitCurve}
\alias{fitCurve}
\title{Curve fitting using minpack.lm}
\usage{
fitCurve(x, y, curveModel = "skewedGaussian", params = "guess",
  lower = NULL, upper = NULL)
}
\arguments{
\item{x}{(numeric) x values (e.g. retention time)}

\item{y}{(numeric) y observed values (e.g. spectra intensity)}

\item{curveModel}{(str) name of the curve model to fit (currently \code{skewedGaussian})}

\item{params}{(list or str) either list of curve fit parameters}

\item{lower}{(NULL or numeric vector) if not NULL, a numeric vector of lower bounds on each parameter. If NULL preset parammeters are employed}

\item{upper}{(NULL or numeric vector) if not NULL, a numeric vector of upper bounds on each parameter. If NULL preset parammeters are employed}
}
\value{
A list of fitted curve parameters, \code{fitStatus} from \code{nls.lm$info} and curve shape name \code{curveModel}. \code{fitStatus=0} unsuccessful completion: improper input parameters, \code{fitStatus=1} successful completion: first convergence test is successful, \code{fitStatus=2} successful completion: second convergence test is successful, \code{fitStatus=3} successful completion: both convergence test are successful, \code{fitStatus=4} questionable completion: third convergence test is successful but should be carefully examined (maximizers and saddle points might satisfy), \code{fitStatus=5} unsuccessful completion: excessive number of function evaluations/iterations
}
\description{
Fit different curve models using minpack. Fitting parameters can be passed or guessed.
}
