% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFit_utils.R
\name{gaussian_erf}
\alias{gaussian_erf}
\title{Gaussian Error function}
\usage{
gaussian_erf(x)
}
\arguments{
\item{x}{(numeric) value at which to evaluate the gaussian error function}
}
\value{
Value of the gaussian error function evaluated at x
}
\description{
Implementation of the gaussian error function
}
