% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcquisitionDatemzML.R
\name{getAcquisitionDatemzML}
\alias{getAcquisitionDatemzML}
\title{Parse acquisition date from a mzML file}
\usage{
getAcquisitionDatemzML(mzMLPath, verbose = TRUE)
}
\arguments{
\item{mzMLPath}{(str) path to mzML raw data file}

\item{verbose}{(bool) if TRUE message progress}
}
\value{
POSIXct or NA
}
\description{
Extract acquisition date (`'startTimeStamp'``) from a mzML file.
In case of failure (or the file is not a \code{mzML}) returns NULL
}
