% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFit_utils.R
\name{is.peakPantheR_curveFit}
\alias{is.peakPantheR_curveFit}
\title{Check if object is of class peakPantheR_curveFit}
\usage{
is.peakPantheR_curveFit(x)
}
\arguments{
\item{x}{object to test}
}
\value{
(bool) TRUE or FALSE
}
\description{
Check if object is of class peakPantheR_curveFit
}
