% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{isAnnotated,peakPantheRAnnotation-method}
\alias{isAnnotated,peakPantheRAnnotation-method}
\alias{isAnnotated}
\title{isAnnotated accessor}
\usage{
\S4method{isAnnotated}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
isAnnotated accessor
}
