% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{nbSamples,peakPantheRAnnotation-method}
\alias{nbSamples,peakPantheRAnnotation-method}
\alias{nbSamples}
\title{nbSamples accessor established on filepath}
\usage{
\S4method{nbSamples}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(int) number of samples
}
\description{
nbSamples accessor established on filepath
}
