% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\name{outputAnnotationParamsCSV}
\alias{outputAnnotationParamsCSV}
\title{Save annotation parameters as CSV}
\usage{
outputAnnotationParamsCSV(annotation, saveFolder, verbose = TRUE)
}
\arguments{
\item{annotation}{(peakPantheRAnnotation) Annotated peakPantheRAnnotation object}

\item{verbose}{(bool) If TRUE message progress}
}
\value{
None
}
\description{
Save annotation parameters (ROI, uROI and FIR) to disk as a CSV file for editing.
}
