% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{peakFit,peakPantheRAnnotation-method}
\alias{peakFit,peakPantheRAnnotation-method}
\alias{peakFit}
\title{peakFit accessor}
\usage{
\S4method{peakFit}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
A list of length number of spectra files. Each list element is a \emph{curveFit} list of \code{peakPantheR_curveFit} or NA for each ROI
}
\description{
peakFit accessor
}
