% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPantheR.R
\name{peakPantheR_findTargetFeature}
\alias{peakPantheR_findTargetFeature}
\title{Integrate ROI and find target features}
\usage{
peakPantheR_findTargetFeature(rawSpec, ROIList, ppm = 20, snthresh = 3,
  noise = 400, prefilter = c(7, 400), peakwidth = c(2, 20),
  verbose = FALSE, fitGauss = FALSE)
}
\arguments{
\item{rawSpec}{an \code{\link[MSnbase]{OnDiskMSnExp}}}

\item{ROIList}{a list of ROIs as generated by \code{\link{peakPantheR_makeROIList}}}

\item{ppm}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{maxmial tolerated m/z deviation in consecutive scans, in ppm (parts per million)}}

\item{snthresh}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{signal to noise ratio cutoff}}

\item{noise}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{optional argument which is useful for data that was centroided without any intensity threshold, centroids with intensity < \code{noise} are omitted from ROI detection}}

\item{prefilter}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{\code{prefilter=c(k,I)}. Prefilter step for the first phase. Mass traces are only retained if they contain at least \code{k} peaks with intensity >= \code{I}.}}

\item{peakwidth}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{Chromatographic peak width, given as range (min,max) in seconds}}

\item{verbose}{(bool) if TRUE message the time taken and number of features found (total and matched to targets)}

\item{fitGauss}{(bool) if TRUE fits peak with option \code{CentWaveParam(..., fitgauss=TRUE)}.}
}
\value{
A \code{data.frame} with targeted features as rows and peak measures as columns (see Details).

\subsection{Details:}{
  The returned \code{data.frame} is structured as follow (from \code{\link[xcms]{findChromPeaks-centWave}}):
  \tabular{ll}{
    mz \tab weighted (by intensity) mean of peak m/z across scans\cr
    mzmin \tab m/z peak minimum\cr
    mzmax \tab m/z peak maximum\cr
    rt \tab retention time of peak midpoint\cr
    rtmin \tab leading edge of peak retention time\cr
    rtmax \tab trailing edge of peak retention time\cr
    into \tab integrated peak intensity\cr
    intb \tab baseline corrected integrated peak intensity\cr
    maxo \tab maximum peak intensity\cr
    sn \tab Signal/Noise ratio, defined as \code{(maxo - baseline)/sd}, where \code{maxo} is the maximum peak intensity, \code{baseline} the estimated baseline value and \code{sd} the standard deviation of local chromatographic noise.\cr
    egauss \tab RMSE of Gaussian fit\cr
    mu \tab Gaussian parameter mu\cr
    sigma \tab Gaussian parameter sigma\cr
    h \tab Gaussian parameter h\cr
    f \tab Region number of m/z ROI where the peak was localised\cr
    dppm \tab m/z deviation of mass trace across scans in ppm\cr
    scale \tab Scale on which the peak was localised\cr
    scpos \tab Peak position found by wavelet analysis (scan number)\cr
    scmin \tab Left peak limit found by wavelet analysis (scan number)\cr
    scmax \tab Right peak limit found by wavelet analysis (scan number)\cr
  }
}
}
\description{
Integrate features in the ROI using \code{CentWave} and keep in each ROI the feature with the highest integrated intensity.
}
\examples{
\donttest{
## Load data
library(MSnbase)
netcdffile <- './my_spectra.CDF'
raw_data   <- MSnbase::readMSData(netcdffile, centroided=TRUE, mode='onDisk')

## targetFeatTable from outside source
targetFeatTable            <- data.frame(matrix(vector(), 11, 8, dimnames=list(c(), c("cpdID", "cpdName", "rtMin", "rt", "rtMax", "mzMin", "mz", "mzMax"))), stringsAsFactors=F)
targetFeatTable[1,]        <- c(1, "LPC (9:0/0:0)", 29.4, NA, 38.4, 398.2108, 398.2308, 398.2508)
targetFeatTable[2,]        <- c(2, "PC (11:0/11:0)", 138.0, NA, 198.0, 594.3935, 594.4135, 594.4335)
targetFeatTable[,c(1,3:8)] <- sapply(targetFeatTable[,c(1,3:8)], as.numeric)

ROIList <- peakPantheR_makeROIList(raw_data, targetFeatTable)

foundFeatTable <- peakPantheR_findTargetFeature(raw_data, ROIList)
foundFeatTable
#   found       mz    mzmin    mzmax      rt   rtmin   rtmax    into    intb      maxo    sn     egauss
# 1  TRUE 398.2319 398.2306 398.2330  33.296  31.468  35.352 1919498 1919222  942711.5 15739 0.04867894
# 2  TRUE 594.4140 594.4120 594.4149 179.046 176.762 181.103 2233011 2232784 1038315.5 30275 0.03296024
#          mu    sigma         h f dppm scale scpos scmin scmax lmin lmax sample is_filled
# 1 136.1303 3.458960  979537.2 1    3     4   136   132   140  128  145      1         0
# 2 770.7220 3.832822 1034072.3 2    1    -1    -1    -1    -1  375  394      1         0
}

}
\seealso{
Other peakPantheR: \code{\link{peakPantheR_cpdSearch}},
  \code{\link{peakPantheR_getTargetFeatureStatistic}},
  \code{\link{peakPantheR_makeROIList}}
}
