% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPantheR_start_GUI.R
\name{peakPantheR_start_GUI}
\alias{peakPantheR_start_GUI}
\title{peakPantheR Graphical User Interface}
\usage{
peakPantheR_start_GUI(browser = TRUE)
}
\arguments{
\item{browser}{If TRUE open the graphical user interface in a web browser instead of a R window. Default is TRUE}
}
\value{
None, start GUI. To exit press \code{ESC} in the command line.
}
\description{
peakPantheR Graphical User Interface (GUI) implements all the functions for the parallel detection, integration and reporting 
of pre-defined features in multiple mass spectrometry data files. To exit press \code{ESC} in the command line.
}
\examples{
\donttest{
## Start graphical interface, press 'ESC' in the command line to stop.
peakPantheR_start_GUI()
}

}
