% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot variable histogram and density}
\usage{
plotHistogram(var, varName = "Variable", density = TRUE, ...)
}
\arguments{
\item{var}{(float) vector of values to plot}

\item{varName}{(str) Name of the variable to plot}

\item{density}{(bool) If TRUE plot overlay the density on the variable}

\item{...}{Passes arguments to ggplot2::geom_histogram, e.g. \code{bins=20}, \code{binwidth=1}}
}
\value{
Grob (ggplot object)
}
\description{
Plot the histogram and density of the variable
}
