% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRunOrder.R
\name{plotRunOrder}
\alias{plotRunOrder}
\title{Plot peak value by acquisition time}
\usage{
plotRunOrder(sampleValue, acquTime, varName = "Variable",
  sampleColour = NULL, verbose = TRUE)
}
\arguments{
\item{sampleValue}{(float) vector of sample value to plot}

\item{acquTime}{(POSIXct) vector of sample acquisition time as POSIXct}

\item{varName}{(str) Name of the variable to plot}

\item{sampleColour}{(str) NULL or vector colour for each sample (same length as \code{sampleValue}, \code{acquTime})}

\item{verbose}{(bool) if TRUE message when NA scans are removed}
}
\value{
Grob (ggplot object)
}
\description{
plot peak value by acquisition time for a single ROI
}
