% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFit_utils.R
\name{predictCurve}
\alias{predictCurve}
\title{Predict curve values}
\usage{
predictCurve(fittedCurve, x)
}
\arguments{
\item{fittedCurve}{(list) a list of curve fitting parameters, curve shape model \code{curveModel} and nls.lm fit status \code{fitStatus}.}

\item{x}{(numeric) values at which to evaluate the fitted curve}
}
\value{
fitted curve values at x
}
\description{
Evaluate fitted curve values at \code{x} data points
}
