% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPantheR_loadAnnotationParamsCSV.R
\name{prepare_advanced_target_parameters}
\alias{prepare_advanced_target_parameters}
\title{Process target region parameters (with uROI, FIR) for object initialisation}
\usage{
prepare_advanced_target_parameters(paramTable, verbose)
}
\arguments{
\item{paramTable}{(data.frame) Target region parameters}

\item{verbose}{(bool) If TRUE message progress}
}
\value{
(list) List of \code{targetFeatTable, uROI, FIR, uROIExist}
}
\description{
Process target region parameters with uROI and FIR (\code{cpdID, cpdName,
ROI_rt, ROI_mz, ROI_rtMin, ROI_rtMax, ROI_mzMin, ROI_mzMax, uROI_rtMin,
uROI_rtMax, uROI_mzMin, uROI_mzMax, uROI_rt, uROI_mz, FIR_rtMin, FIR_rtMax,
FIR_mzMin, FIR_mzMax}) and return input variables for
\code{peakPantheRAnnotation()}
}
