% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{resetFIR,peakPantheRAnnotation-method}
\alias{resetFIR,peakPantheRAnnotation-method}
\alias{resetFIR}
\title{Reset FIR windows to uROI or ROI values
Reset FIR windows to uROI (or ROI if \code{uROIExist=FALSE})}
\usage{
\S4method{resetFIR}{peakPantheRAnnotation}(object, verbose)
}
\arguments{
\item{object}{(peakPantheRAnnotation) object for which FIR are to be reset}

\item{verbose}{(bool) If TRUE message progress}
}
\value{
(peakPantheRAnnotation) object with FIR values reset
}
\description{
Reset FIR windows to uROI or ROI values
Reset FIR windows to uROI (or ROI if \code{uROIExist=FALSE})
}
