% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveFit_utils.R
\name{skewedGaussian_guess}
\alias{skewedGaussian_guess}
\title{Guess function for initial skewed gaussian parameters}
\usage{
skewedGaussian_guess(x, y)
}
\arguments{
\item{x}{(numeric) x values (e.g. retention time)}

\item{y}{(numeric) y observed values (e.g. spectra intensity)}
}
\value{
Guessed starting parameters (\code{params$gamma}, \code{params$center}, \code{params$sigma}, \code{params$amplitude})
}
\description{
Guess function for initial skewed gaussian parameters, at the moment only checks the x position
}
