% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{spectraMetadata,peakPantheRAnnotation-method}
\alias{spectraMetadata,peakPantheRAnnotation-method}
\alias{spectraMetadata}
\title{spectraMetadata accessor}
\usage{
\S4method{spectraMetadata}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
spectraMetadata accessor
}
