% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{[,peakPantheRAnnotation,ANY,ANY,ANY-method}
\alias{[,peakPantheRAnnotation,ANY,ANY,ANY-method}
\alias{[,peakPantheRAnnotation-method}
\title{extract parts of peakPantheRAnnotation class}
\usage{
\S4method{[}{peakPantheRAnnotation,ANY,ANY,ANY}(x, i, j, drop = "missing")
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace
element(s).}

\item{i}{(sample) indices specifying elements to extract or replace}

\item{j}{(compound) indices specifying elements to extract or replace}

\item{drop}{not applicable}
}
\value{
(peakPantheRAnnotation) object subset
}
\description{
extract parts of peakPantheRAnnotation class
}
