% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{TIC,peakPantheRAnnotation-method}
\alias{TIC,peakPantheRAnnotation-method}
\alias{TIC}
\title{TIC accessor}
\usage{
\S4method{TIC}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(float) A numeric vector of Total Ion Chromatogram or NA, of length number of spectra files
}
\description{
TIC accessor
}
