% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{uROIExist,peakPantheRAnnotation-method}
\alias{uROIExist,peakPantheRAnnotation-method}
\alias{uROIExist}
\title{uROIExist accessor}
\usage{
\S4method{uROIExist}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(bool) flag if uROI have been set
}
\description{
uROIExist accessor
}
