% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{useFIR,peakPantheRAnnotation-method}
\alias{useFIR,peakPantheRAnnotation-method}
\alias{useFIR}
\title{useFIR accessor}
\usage{
\S4method{useFIR}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(bool) flag if FIR are to be used
}
\description{
useFIR accessor
}
