% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{useUROI,peakPantheRAnnotation-method}
\alias{useUROI,peakPantheRAnnotation-method}
\alias{useUROI}
\title{useUROI accessor}
\usage{
\S4method{useUROI}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(bool) flag if uROI are to be used
}
\description{
useUROI accessor
}
