% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{annotation_diagnostic_multiplot_UI_helper}
\alias{annotation_diagnostic_multiplot_UI_helper}
\title{UI diagnostic plot helper - single feature multiplot}
\usage{
annotation_diagnostic_multiplot_UI_helper(
    cpdNb,
    annotation,
    sampleColour = NULL,
    ...
)
}
\arguments{
\item{cpdNb}{(int) postion of the feature to extract (1 to nbCpd)}

\item{annotation}{(peakPantheRAnnotation) Annotation object}

\item{...}{Additional parameters for plotting}

\item{sampleColourColumn}{(str) NULL, None or a spectraMetadata column for
colouring each sample}
}
\value{
(ggplotObject) Diagnostic multiplot for a feature
}
\description{
Return a ggplot object of a feature diagnostic multiplot
}
