% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{annotation_diagnostic_multiplot_UI_helper}
\alias{annotation_diagnostic_multiplot_UI_helper}
\title{UI diagnostic plot helper - single feature multiplot}
\usage{
annotation_diagnostic_multiplot_UI_helper(
    cpdNb,
    annotation,
    splNum = NULL,
    splColrColumn = NULL,
    ...
)
}
\arguments{
\item{cpdNb}{(int) position of the feature to extract (1 to nbCpd)}

\item{annotation}{(peakPantheRAnnotation) Annotation object}

\item{splNum}{(int) NULL or number of spectra to plot, chosen randomly
from all spectra. If NULL or equal to the total number of spectra, plot all
spectra}

\item{splColrColumn}{(str) NULL, None or a spectraMetadata column for
colouring each sample}

\item{...}{Additional parameters for plotting}
}
\value{
(ggplotObject) Diagnostic multiplot for a feature
}
\description{
Return a ggplot object of a feature diagnostic multiplot
}
