% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{annotation_showMethod_UI_helper}
\alias{annotation_showMethod_UI_helper}
\title{UI show annotation helper - list of properties}
\usage{
annotation_showMethod_UI_helper(annotation)
}
\arguments{
\item{annotation}{(peakPantherAnnotation) Object to describe}
}
\value{
(list) Named list of annotation properties
}
\description{
show method specific to the UI, that returns each field in a named list to
ease display
}
\examples{
# Initialise an empty annotation, no uROI, no use of FIR
annotInit     <- peakPantheRAnnotation()

# return properties
annotation_showMethod_UI_helper(annotInit)
# $nbCompounds
# [1] 0
#
# #$nbSamples
# [1] 0
#
# #$uROIExist
# [1] FALSE
#
# $useUROI
# [1] FALSE
#
# $useFIR
# [1] FALSE
#
# $isAnnotated
# [1] FALSE
}
