% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{annotation_showText_UI_helper}
\alias{annotation_showText_UI_helper}
\title{UI show annotation helper - UI sidebar string}
\usage{
annotation_showText_UI_helper(annotProp)
}
\arguments{
\item{annotProp}{(list) Named list of annotation properties as
created by \code{annotation_showMethod_UI_helper()}}
}
\value{
(str) Textual description of the annotation to show on UI
}
\description{
Return a text description of an annotation for UI presentation
}
