% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{cpdID,peakPantheRAnnotation-method}
\alias{cpdID,peakPantheRAnnotation-method}
\alias{cpdID}
\title{cpdID accessor}
\usage{
\S4method{cpdID}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(str) A character vector of compound IDs, of length number of compounds
}
\description{
cpdID accessor
}
