% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{dataPoints,peakPantheRAnnotation-method}
\alias{dataPoints,peakPantheRAnnotation-method}
\alias{dataPoints}
\title{dataPoints accessor}
\usage{
\S4method{dataPoints}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
dataPoints accessor
}
