% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{EICs,peakPantheRAnnotation-method}
\alias{EICs,peakPantheRAnnotation-method}
\alias{EICs}
\title{EICs accessor}
\usage{
\S4method{EICs}{peakPantheRAnnotation}(object, aggregationFunction)
}
\arguments{
\item{object}{peakPantheRAnnotation}

\item{aggregationFunction}{(str) Function to use in order to aggregate intensities across mz in each scan. One of \code{sum}, \code{max}, \code{min}, \code{mean}}
}
\description{
EICs accessor
}
