% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{filepath,peakPantheRAnnotation-method}
\alias{filepath,peakPantheRAnnotation-method}
\alias{filepath}
\title{filepath accessor}
\usage{
\S4method{filepath}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(str) A character vector of file paths, of length number of spectra files
}
\description{
filepath accessor
}
