% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findTargetFeatures}
\alias{findTargetFeatures}
\title{Integrate ROI and find target features}
\usage{
findTargetFeatures(rawSpec, ROIList, ppm = 20, snthresh = 3, noise = 400,
  prefilter = c(7, 400), peakwidth = c(2, 20), verbose = FALSE,
  fitGauss = FALSE)
}
\arguments{
\item{rawSpec}{an \code{\link[MSnbase]{OnDiskMSnExp-class}}}

\item{ROIList}{a list of ROIs as generated by \code{\link{makeROIList}}}

\item{ppm}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{maxmial tolerated m/z deviation in consecutive scans, in ppm (parts per million)}}

\item{snthresh}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{signal to noise ratio cutoff}}

\item{noise}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{optional argument which is useful for data that was centroided without any intensity threshold, centroids with intensity < \code{noise} are omitted from ROI detection}}

\item{prefilter}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{\code{prefilter=c(k,I)}. Prefilter step for the first phase. Mass traces are only retained if they contain at least \code{k} peaks with intensity >= \code{I}.}}

\item{peakwidth}{\code{\link[xcms]{findChromPeaks-centWave}} parameter: \emph{Chromatographic peak width, given as range (min,max) in seconds}}

\item{verbose}{(bool) if TRUE message the time taken and number of features found (total and matched to targets)}

\item{fitGauss}{(bool) if TRUE fits peak with option \code{CentWaveParam(..., fitgauss=TRUE)}.}
}
\value{
A \code{data.frame} with targeted features as rows and peak measures as columns (see Details).

\subsection{Details:}{
  The returned \code{data.frame} is structured as follow (from \code{\link[xcms]{findChromPeaks-centWave}}):
  \tabular{ll}{
    mz \tab weighted (by intensity) mean of peak m/z across scans\cr
    mzmin \tab m/z peak minimum\cr
    mzmax \tab m/z peak maximum\cr
    rt \tab retention time of peak midpoint\cr
    rtmin \tab leading edge of peak retention time\cr
    rtmax \tab trailing edge of peak retention time\cr
    into \tab integrated peak intensity\cr
    intb \tab baseline corrected integrated peak intensity\cr
    maxo \tab maximum peak intensity\cr
    sn \tab Signal/Noise ratio, defined as \code{(maxo - baseline)/sd}, where \code{maxo} is the maximum peak intensity, \code{baseline} the estimated baseline value and \code{sd} the standard deviation of local chromatographic noise.\cr
    egauss \tab RMSE of Gaussian fit\cr
    mu \tab Gaussian parameter mu\cr
    sigma \tab Gaussian parameter sigma\cr
    h \tab Gaussian parameter h\cr
    f \tab Region number of m/z ROI where the peak was localised\cr
    dppm \tab m/z deviation of mass trace across scans in ppm\cr
    scale \tab Scale on which the peak was localised\cr
    scpos \tab Peak position found by wavelet analysis (scan number)\cr
    scmin \tab Left peak limit found by wavelet analysis (scan number)\cr
    scmax \tab Right peak limit found by wavelet analysis (scan number)\cr
  }
}
}
\description{
Integrate features in the ROI using \code{CentWave} and keep in each ROI the feature with the highest integrated intensity.
}
\examples{
\dontrun{
## Load data
library(faahKO)
library(MSnbase)
netcdfFilePath <- system.file('cdf/KO/ko15.CDF', package = "faahKO")
raw_data       <- MSnbase::readMSData(netcdfFilePath, centroided=TRUE, mode='onDisk')

## targetFeatTable
targetFeatTable     <- data.frame(matrix(vector(), 2, 8, dimnames=list(c(), c("cpdID",
                         "cpdName", "rtMin", "rt", "rtMax", "mzMin", "mz", "mzMax"))),
                         stringsAsFactors=F)
targetFeatTable[1,] <- c(1, "Cpd 1", 3310., 3344.888, 3390., 522.194778, 522.2, 522.205222)
targetFeatTable[2,] <- c(2, "Cpd 2", 3280., 3385.577, 3440., 496.195038, 496.2, 496.204962)
targetFeatTable[,c(1,3:8)] <- sapply(targetFeatTable[,c(1,3:8)], as.numeric)

ROIList        <- makeROIList(raw_data, targetFeatTable)

foundPeakTable <- findTargetFeatures(raw_data, ROIList)
foundPeakTable
#   found    mz mzmin mzmax       rt    rtmin    rtmax     into     intb    maxo     sn
# 1  TRUE 522.2 522.2 522.2 3344.888 3322.979 3379.317 25792525 25768308  889280   1840
# 2  TRUE 496.2 496.2 496.2 3382.447 3362.102 3409.051 32873727 32818664 1128960   1471
# egauss mu sigma  h f dppm scale scpos scmin scmax lmin lmax sample is_filled
#     NA NA    NA NA 1    0     5   540   535   545   24   60      1         0
#     NA NA    NA NA 2    0     5   564   559   569   68   98      1         0
}
}
