% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{FIR,peakPantheRAnnotation-method}
\alias{FIR,peakPantheRAnnotation-method}
\alias{FIR}
\title{FIR accessor returns targetFeatTable with cpdID, cpdName added}
\usage{
\S4method{FIR}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(data.frame) target feature table with compounds as row and FIR parameters as columns
}
\description{
FIR accessor returns targetFeatTable with cpdID, cpdName added
}
