% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargetFeatureStatistic.R
\name{getTargetFeatureStatistic}
\alias{getTargetFeatureStatistic}
\title{Calculate chromatographic peak properties}
\usage{
getTargetFeatureStatistic(fittedCurves, targetFeatTable, foundPeakTable,
  verbose = FALSE)
}
\arguments{
\item{fittedCurves}{(list) A list (one entry per ROI window) of \code{peakPantheR_curveFit} or NA}

\item{targetFeatTable}{a \code{\link{data.frame}} of compounds to target as rows. Columns: \code{cpdID} (str), \code{cpdName} (str), \code{rtMin} (float in seconds), \code{rt} (float in seconds, or \emph{NA}), \code{rtMax} (float in seconds), \code{mzMin} (float), \code{mz} (float or \emph{NA}), \code{mzMax} (float).}

\item{foundPeakTable}{a \code{data.frame} as generated by \code{\link{findTargetFeatures}}, with features as rows and peak properties as columns. The following columns are mandatory: \code{mzMin}, \code{mz}, \code{mzMax}, \code{rtMin}, \code{rt}, \code{rtMax}}

\item{verbose}{(bool) if TRUE message when NA scans are removed}
}
\value{
A \code{data.frame} with measured compounds as rows and measurements and properties as columns (see Details).

\subsection{Details:}{
  The returned \code{data.frame} is structured as follow:
  \tabular{ll}{
    found \tab was the peak found\cr
    rt \tab retention time of peak apex (sec)\cr
    rtMin \tab leading edge of peak retention time (sec) determined at 0.5\% of apex intensity\cr
    rtMax \tab trailing edge of peak retention time (sec) determined at 0.5\% of apex intensity\cr
    mz \tab weighted (by intensity) mean of peak m/z across scans\cr
    mzMin \tab m/z peak minimum (between rtMin, rtMax)\cr
    mzMax \tab m/z peak maximum (between rtMin, rtMax)\cr
    peakArea \tab integrated peak area\cr
    maxIntMeasured \tab maximum peak intensity in raw data\cr
    maxIntPredicted \tab maximum peak intensity based on curve fit\cr
    ppm_error \tab difference in ppm between the expected and measured m/z\cr
    rt_dev_sec \tab difference in seconds between the expected and measured rt\cr
    tailingFactor \tab the tailing factor is a measure of peak tailing. It is defined as the distance from the front slope of the peak to the back slope divided by twice the distance from the center line of the peak to the front slope, with all measurements made at 5\% of the maximum peak height. The tailing factor of a peak will typically be similar to the asymmetry factor for the same peak, but the two values cannot be directly converted\cr
   asymmetryFactor \tab the asymmetry factor is a measure of peak tailing. It is defined as the distance from the center line of the peak to the back slope divided by the distance from the center line of the peak to the front slope, with all measurements made at 10\% of the maximum peak height. The asymmetry factor of a peak will typically be similar to the tailing factor for the same peak, but the two values cannot be directly converted\cr
  }
}
}
\description{
Calculate the ppm error, retention time deviation, tailing factor and asymmetry factor for each measured feature.
}
\examples{
\dontrun{
## Load data
library(faahKO)
library(MSnbase)
netcdfFilePath <- system.file('cdf/KO/ko15.CDF', package = "faahKO")
raw_data       <- MSnbase::readMSData(netcdfFilePath, centroided=TRUE, mode='onDisk')

## targetFeatTable
targetFeatTable     <- data.frame(matrix(vector(), 2, 8, dimnames=list(c(), c("cpdID",
                         "cpdName", "rtMin", "rt", "rtMax", "mzMin", "mz", "mzMax"))),
                         stringsAsFactors=F)
targetFeatTable[1,] <- c("ID-1", "Cpd 1", 3310., 3344.888, 3390., 522.194778, 522.2, 522.205222)
targetFeatTable[2,] <- c("ID-2", "Cpd 2", 3280., 3385.577, 3440., 496.195038, 496.2, 496.204962)
targetFeatTable[,c(1,3:8)] <- sapply(targetFeatTable[,c(1,3:8)], as.numeric)

ROIList        <- makeROIList(raw_data, targetFeatTable)

foundPeakTable <- findTargetFeatures(raw_data, ROIList)

peakStatistics <- getTargetFeatureStatistic(raw_data, targetFeatTable, foundPeakTable)
peakStatistics
#    found    mz mzmin mzmax       rt    rtmin    rtmax     into     intb   maxo    sn egauss mu
# 1   TRUE 522.2 522.2 522.2 3344.888 3322.979 3379.317 25792525 25768308 889280  1840     NA NA
# 2   TRUE 496.2 496.2 496.2 3382.447 3362.102 3409.051 32873727 32818664 1128960 1471     NA NA
#    sigma h  f dppm scale scpos scmin scmax lmin lmax sample is_filled   ppm_error rt_dev_sec
# 1     NA NA 1    0     5   540   535   545   24   60      1         0  0.02336270       0.00
# 2     NA NA 2    0     5   564   559   569   68   98      1         0  0.02458686       3.13
#    FWHM FWHM_ndatapoints tailingFactor asymmetryFactor
# 1    NA               11            NA        1.484000
# 2    NA               11            NA        2.708291
}
}
