% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTargetFeatureStatistic}
\alias{getTargetFeatureStatistic}
\title{Calculate chromatographic peak properties}
\usage{
getTargetFeatureStatistic(rawSpec, targetFeatTable, foundFeatTable)
}
\arguments{
\item{rawSpec}{an \code{\link[MSnbase]{OnDiskMSnExp}}}

\item{targetFeatTable}{a \code{\link{data.frame}} of compounds to target as rows. Columns: \code{cpdID} (int), \code{cpdName} (str), \code{rtMin} (float in seconds), \code{rt} (float in seconds, or \emph{NA}), \code{rtMax} (double in seconds), \code{mzMin} (float in seconds), \code{mz} (float in seconds, or \emph{NA}), \code{mzMax} (float in seconds).}

\item{foundFeatTable}{a \code{data.frame} as generated by \code{\link{findTargetFeatures}}, with features as rows and peak properties as columns. The following columns are mandatory: \code{mz}, \code{rt}, \code{scmin}, \code{scpos}, \code{scmax}, \code{sigma}.}
}
\value{
A \code{data.frame} with measured compounds as rows and measurements and properties as columns (see Details).

\subsection{Details:}{
  The returned \code{data.frame} is structured as follow:
  \tabular{ll}{
    cpdID \tab database compound ID\cr
    cpdName \tab compound name\cr
    found \tab (bool) TRUE if compound was found in the raw data\cr
    mz \tab weighted (by intensity) mean of peak m/z across scans\cr
    mzmin \tab m/z peak minimum\cr
    mzmax \tab m/z peak maximum\cr
    rt \tab retention time of peak midpoint\cr
    rtmin \tab leading edge of peak retention time\cr
    rtmax \tab trailing edge of peak retention time\cr
    into \tab integrated peak intensity\cr
    intb \tab baseline corrected integrated peak intensity\cr
    maxo \tab maximum peak intensity\cr
    sn \tab Signal/Noise ratio, defined as \code{(maxo - baseline)/sd}, where \code{maxo} is the maximum peak intensity, \code{baseline} the estimated baseline value and \code{sd} the standard deviation of local chromatographic noise.\cr
    egauss \tab RMSE of Gaussian fit\cr
    mu \tab Gaussian parameter mu\cr
    sigma \tab Gaussian parameter sigma\cr
    h \tab Gaussian parameter h\cr
    f \tab Region number of m/z ROI where the peak was localised\cr
    dppm \tab m/z deviation of mass trace across scans in ppm\cr
    scale \tab Scale on which the peak was localised\cr
    scpos \tab Peak position found by wavelet analysis (scan number)\cr
    scmin \tab Left peak limit found by wavelet analysis (scan number)\cr
    scmax \tab Right peak limit found by wavelet analysis (scan number)\cr
    ppm_error \tab difference in ppm between the expected and measured m/z\cr
    rt_dev_sec \tab difference in seconds between the expected and measured rt\cr
    FWHM \tab full width at half maximum (in seconds), not available if \code{fitGauss=FALSE}\cr
    FWHM_ndatapoints \tab number of scans on the peak\cr
    tailingFactor \tab the tailing factor is a measure of peak tailing.It is defined as the distance from the front slope of the peak to the back slope divided by twice the distance from the center line of the peak to the front slope, with all measurements made at 5\% of the maximum peak height. The tailing factor of a peak will typically be similar to the asymmetry factor for the same peak, but the two values cannot be directly converted\cr
    assymmetryFactor \tab the asymmetry factor is a measure of peak tailing. It is defined as the distance from the center line of the peak to the back slope divided by the distance from the center line of the peak to the front slope, with all measurements made at 10\% of the maximum peak height. The asymmetry factor of a peak will typically be similar to the tailing factor for the same peak, but the two values cannot be directly converted\cr
  }
}
}
\description{
Calculate the ppm error, retention time deviation, FWHM, number of scans, tailing factor and assymmetry factor for each measured feature.
}
\examples{
\donttest{
## Load data
library(MSnbase)
netcdfFilePath <- './my_spectra.CDF'
raw_data       <- MSnbase::readMSData(netcdfFilePath, centroided=TRUE, mode='onDisk')

## targetFeatTable
targetFeatTable     <- data.frame(matrix(vector(), 11, 8, dimnames=list(c(), c("cpdID",
                         "cpdName", "rtMin", "rt", "rtMax", "mzMin", "mz", "mzMax"))),
                         stringsAsFactors=F)
targetFeatTable[1,] <- c(1, "LPC (9:0/0:0)", 29.4, NA, 38.4, 398.2108, 398.2308, 398.2508)
targetFeatTable[2,] <- c(2, "PC (11:0/11:0)", 138.0, NA, 198.0, 594.3935, 594.4135, 594.4335)
targetFeatTable[,c(1,3:8)] <- sapply(targetFeatTable[,c(1,3:8)], as.numeric)

ROIList        <- makeROIList(raw_data, targetFeatTable)

foundFeatTable <- findTargetFeatures(raw_data, ROIList)

finalOutput    <- getTargetFeatureStatistic(raw_data, targetFeatTable, foundFeatTable)
finalOutput
#    cpdID        cpdName found       mz    mzmin    mzmax      rt   rtmin   rtmax    into    intb
# 1      1  LPC (9:0/0:0)  TRUE 398.2319 398.2306 398.2330  33.296  31.468  35.352 1919498 1919222
# 2      2 PC (11:0/11:0)  TRUE 594.4140 594.4120 594.4149 179.046 176.762 181.103 2233011 2232784
#         maxo    sn     egauss       mu    sigma         h f dppm  scale scpos scmin scmax lmin
# 1   942711.5 15739 0.04867894 136.1303 3.458960  979537.2 1    3      4   136   132   140  128
# 2  1038315.5 30275 0.03296024 770.7220 3.832822 1034072.3 2    1     -1    -1    -1    -1  375
#    lmax sample is_filled ppm_error rt_dev_sec    FWHM FWHM_ndatapoints tailingFactor
# 1   145      1         0  2.744491         NA 1.86026                9     0.8868141
# 2   394      1         0  0.830051         NA      NA               NA     1.0053721
#    assymmetryFactor
# 1         1.1248373
# 2         0.9744952
}

}
\references{
Adapted for XCMS3 from Jan Stanstrup https://cdn.rawgit.com/stanstrup/QC4Metabolomics/master/MetabolomiQCsR/inst/doc/standard_stats.html#calculating-statistics
}
\seealso{
Other peakPantheR: \code{\link{findTargetFeatures}},
  \code{\link{makeROIList}},
  \code{\link{peakPantheR_singleFileSearch}}
}
