% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{load_annotation_from_file_UI_helper}
\alias{load_annotation_from_file_UI_helper}
\title{UI data import helper - check loaded annotation}
\usage{
load_annotation_from_file_UI_helper(annotationPath)
}
\arguments{
\item{annotationPath}{(str) Path to a RData file containing a
peakPantheRAnnotation names `annotationObject`}
}
\value{
(peakPantheRAnnotation) Object loaded from file
}
\description{
Load a .RData file (check it exists) and that a peakPantheRAnnotation named
"annotationObject" is present. Returns the annotation if everything is valid
}
