% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeROIList}
\alias{makeROIList}
\title{Generate a Region Of Interest (ROI) List}
\usage{
makeROIList(rawSpec, targetFeatTable)
}
\arguments{
\item{rawSpec}{an \code{\link[MSnbase]{OnDiskMSnExp-class}} used to get the scans corresponding to each retention time.}

\item{targetFeatTable}{a \code{\link{data.frame}} of compounds to target as rows. Columns: \code{cpdID} (int), \code{cpdName} (str), \code{rtMin} (float in seconds), \code{rt} (float in seconds, or \emph{NA}), \code{rtMax} (double in seconds), \code{mzMin} (float in seconds), \code{mz} (float in seconds, or \emph{NA}), \code{mzMax} (float in seconds).}
}
\value{
a list of ROIs
}
\description{
Generate a ROIList as expected by \code{\link[xcms]{findChromPeaks-centWave}} from a \code{\link{data.frame}} with compounds to target as rows. \emph{length} and \emph{intensity} are set to \code{-1} as centWave does not use these values.
}
\examples{
\donttest{
## Load data
library(MSnbase)
netcdfFilePath <- './my_spectra.CDF'
raw_data       <- MSnbase::readMSData(netcdfFilePath, centroided=TRUE, mode='onDisk')

## targetFeatTable
targetFeatTable     <- data.frame(matrix(vector(), 11, 8, dimnames=list(c(), c("cpdID",
                         "cpdName", "rtMin", "rt", "rtMax", "mzMin", "mz", "mzMax"))),
                         stringsAsFactors=F)
targetFeatTable[1,] <- c(1, "LPC (9:0/0:0)", 29.4, NA, 38.4, 398.2108, 398.2308, 398.2508)
targetFeatTable[2,] <- c(2, "PC (11:0/11:0)", 138.0, NA, 198.0, 594.3935, 594.4135, 594.4335)
targetFeatTable[,c(1,3:8)] <- sapply(targetFeatTable[,c(1,3:8)], as.numeric)

ROIList <- makeROIList(raw_data, targetFeatTable)
ROIList[[1]]
# $mz
# [1] 398.2308
#
# $mzmin
# [1] 398.2108
#
# $mzmax
# [1] 398.2508
#
# $scmin
# [1] 119
#
# $scmax
# [1] 158
#
# $length
# [1] -1
#
# $intensity
# [1] -1
}
}
