% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{nbCompounds,peakPantheRAnnotation-method}
\alias{nbCompounds,peakPantheRAnnotation-method}
\alias{nbCompounds}
\title{nbCompounds accessor established on cpdID}
\usage{
\S4method{nbCompounds}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
nbCompounds accessor established on cpdID
}
