% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{outputAnnotationFeatureMetadata_UI_helper}
\alias{outputAnnotationFeatureMetadata_UI_helper}
\title{UI export helper - feature metadata}
\usage{
outputAnnotationFeatureMetadata_UI_helper(annot)
}
\arguments{
\item{annot}{(peakPantheRAnnotation) Annotation object}
}
\value{
(data.frame) Features metadata
}
\description{
Return a table with features as rows and all feature metadata as columns
}
