% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{outputAnnotationParamsCSV,peakPantheRAnnotation-method}
\alias{outputAnnotationParamsCSV,peakPantheRAnnotation-method}
\alias{outputAnnotationParamsCSV}
\title{Save annotation parameters as CSV}
\usage{
\S4method{outputAnnotationParamsCSV}{peakPantheRAnnotation}(
    object,
    saveFolder,
    verbose,
    noSave
)
}
\arguments{
\item{object}{(peakPantheRAnnotation) Annotated peakPantheRAnnotation object}

\item{saveFolder}{(str) Path of folder where annotationParameters_summary.csv
will be saved}

\item{verbose}{(bool) If TRUE message progress}

\item{noSave}{(bool) If TRUE the resulting table will be returned without
saving to disk}
}
\value{
None
}
\description{
Save annotation parameters (ROI, uROI and FIR) to disk as a CSV
file for editing
}
