% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{outputAnnotationResult,peakPantheRAnnotation-method}
\alias{outputAnnotationResult,peakPantheRAnnotation-method}
\alias{outputAnnotationResult}
\title{Save to disk all annotation results as csv files
Save to disk all annotation results as \code{annotationName_ ... .csv} files: compound metadata (\code{cpdMetadata}, \code{cpdID}, \code{cpdName}) and spectra metadata (\code{spectraMetadata}, \code{acquisitionTime}, \code{TIC}) and a file for each column of \code{peakTables} (with samples as rows and compounds as columns)}
\usage{
\S4method{outputAnnotationResult}{peakPantheRAnnotation}(object, saveFolder,
  annotationName = "annotationResult", verbose = TRUE)
}
\arguments{
\item{object}{(peakPantheRAnnotation) Annotated peakPantheRAnnotation object}

\item{saveFolder}{(str) Path of folder where the annotation result csv will be saved}

\item{annotationName}{(str) name of annotation to use in the saved csv}

\item{verbose}{(bool) If TRUE message progress}
}
\value{
None
}
\description{
Save to disk all annotation results as csv files
Save to disk all annotation results as \code{annotationName_ ... .csv} files: compound metadata (\code{cpdMetadata}, \code{cpdID}, \code{cpdName}) and spectra metadata (\code{spectraMetadata}, \code{acquisitionTime}, \code{TIC}) and a file for each column of \code{peakTables} (with samples as rows and compounds as columns)
}
