% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{outputAnnotationSpectraMetadata_UI_helper}
\alias{outputAnnotationSpectraMetadata_UI_helper}
\title{UI export helper - spectra path and metadata}
\usage{
outputAnnotationSpectraMetadata_UI_helper(annot)
}
\arguments{
\item{annot}{(peakPantheRAnnotation) Annotation object}
}
\value{
(data.frame) Spectra paths and metadata
}
\description{
Return a table with spectra as rows and filepath and all spectra metadata
columns
}
