% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{peakFit,peakPantheRAnnotation-method}
\alias{peakFit,peakPantheRAnnotation-method}
\alias{peakFit}
\title{peakFit accessor}
\usage{
\S4method{peakFit}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\description{
peakFit accessor
}
