% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPantheR_parallelAnnotation.R
\name{peakPantheR_parallelAnnotation}
\alias{peakPantheR_parallelAnnotation}
\title{Search, integrate and report targeted features in a multiple spectra}
\usage{
peakPantheR_parallelAnnotation(object, ncores = 0, fitGauss = FALSE,
  getAcquTime = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{(peakPantheRAnnotation) Initialised peakPantheRAnnotation object defining the samples to process and compounds to target. The slots \code{useUROI} and \code{useFIR} controls if uROI must be used and FIR integrated if a feature is not found}

\item{ncores}{(int) Number of cores to use for parallelisation. Default 0 for no parallelisation.}

\item{fitGauss}{(bool) if TRUE fits peak with option \code{CentWaveParam(..., fitgauss=TRUE)}.}

\item{getAcquTime}{(bool) If TRUE will extract sample acquisition date-time from the mzML metadata (the additional file access will impact run time)}

\item{verbose}{(bool) If TRUE message calculation progress, time taken, number of features found (total and matched to targets) and failures}

\item{...}{Passes arguments to \code{findTargetFeatures} to alter peak-picking parameters}
}
\value{
a list: \code{list()$result} \emph{(peakPantheRAnnotation)} fully annotated object, \code{list()$failures} \emph{(list)} list of failed samples and error message
}
\description{
Integrate all target features in all files defined in the initialised input object and store results. The use of updated ROI and the integration of FIR are controled by the input object slots \code{useUROI} and \code{useFIR}. Files are processed in parallel using \code{link{peakPantheR_singleFileSearch}}; \code{ncores} controls the number of cores used for parallelisation, with \code{ncores=0} corresponding to serial processing. If the processing of a file fails (file does not exist or error during execution) the sample is removed from the outputed object.
}
\examples{


}
\seealso{
Other peakPantheR: \code{\link{peakPantheRAnnotation}},
  \code{\link{peakPantheR_singleFileSearch}}

Other parallelAnnotation: \code{\link{peakPantheRAnnotation}},
  \code{\link{peakPantheR_singleFileSearch}}
}
