% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPantheR_ROIStatistics.R
\name{peakPantheR_ROIStatistics}
\alias{peakPantheR_ROIStatistics}
\title{Save to disk each ROI EIC and mean IS RT
Using reference samples (\code{referenceSpectraFiles}), save (to \code{saveFolder}) each ROI EIC (\code{ROI}) and reports the mean apex RT for all IS (\code{IS_ROI}) across samples}
\usage{
peakPantheR_ROIStatistics(referenceSpectraFiles, saveFolder, ROI = NULL,
  IS_ROI = NULL, sampleColour = NULL, ncores = 0, saveISPlots = TRUE,
  verbose = TRUE)
}
\arguments{
\item{referenceSpectraFiles}{(str) A character vector of paths to the reference spectra files}

\item{saveFolder}{(str) Path to the folder where EICs and IS mean RT (\code{IS_mean_RT.csv}) will be saved}

\item{ROI}{(data.frame) NULL or a data.frame of Regions Of Interest (ROI) with compounds as row and ROI parameters as columns: \code{rtMin} (float in seconds), \code{rt} (float in seconds, or \emph{NA}), \code{rtMax} (float in seconds), \code{mzMin} (float), \code{mz} (float or \emph{NA}), \code{mzMax} (float) (if NULL, ROI EICs are not saved)}

\item{IS_ROI}{(data.frame) NULL or a data.frame of IS ROI with IS as row and ROI parameters as columns: \code{rtMin} (float in seconds), \code{rt} (float in seconds, or \emph{NA}), \code{rtMax} (float in seconds), \code{mzMin} (float), \code{mz} (float or \emph{NA}), \code{mzMax} (float)  (if NULL IS mean RT is not calculated and saved in \code{IS_mean_RT.csv})}

\item{sampleColour}{(str) NULL or vector colour for each sample}

\item{ncores}{(int) Number of cores to use to integrate IS in parallel}

\item{saveISPlots}{(bool) If TRUE save a diagnostic plot for each IS to \code{saveFolder/IS_search} compound}

\item{verbose}{(bool) If TRUE message progress}
}
\value{
None
}
\description{
Save to disk each ROI EIC and mean IS RT
Using reference samples (\code{referenceSpectraFiles}), save (to \code{saveFolder}) each ROI EIC (\code{ROI}) and reports the mean apex RT for all IS (\code{IS_ROI}) across samples
}
