% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{peakTables,peakPantheRAnnotation-method}
\alias{peakTables,peakPantheRAnnotation-method}
\alias{peakTables}
\title{peakTables accessor with cpdID and cpdName added back}
\usage{
\S4method{peakTables}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(data.frame) A list of peakTable data.frame, of length number of spectra files. Each peakTable data.frame has compounds as rows and peak annotation results as columns, with added compound ID and name.
}
\description{
peakTables accessor with cpdID and cpdName added back
}
