% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotEICDetectedPeakwidth}
\alias{plotEICDetectedPeakwidth}
\title{Plot a raw data and detected feature in a ROI}
\usage{
plotEICDetectedPeakwidth(EIC, cpdID, cpdName, rt, rtmin, rtmax, mzmin, mzmax)
}
\arguments{
\item{EIC}{xcms::chromatograms (single spectra)}

\item{cpdID}{(str) Compound ID}

\item{cpdName}{(str) Compound Name}

\item{rt}{(float) detected peak apex retention time (in sec)}

\item{rtmin}{(float) detected peak minimum retention time (in sec)}

\item{rtmax}{(float) detected peak maximum retention time (in sec)}

\item{mzmin}{(float) ROI minimum m/z (matching EIC)}

\item{mzmax}{(float) ROI maximum m/z (matching EIC)}
}
\value{
Grob (ggplot object)
}
\description{
plot a ROI (x is RT, y is intensity) with the matching detected peak rt and peakwidth under it
}
