% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeakwidth.R
\name{plotPeakwidth}
\alias{plotPeakwidth}
\title{Plot detected peaks width for a single ROI}
\usage{
plotPeakwidth(apexValue, widthMin, widthMax, varName = "Variable",
  sampleColour = NULL, verbose = TRUE)
}
\arguments{
\item{apexValue}{(float) vector of detected peak apex value}

\item{widthMin}{(float) vector of detected peak minimum peakwidth value}

\item{widthMax}{(float) vector of detected peak maximum peakwidth value}

\item{varName}{(str) Name of the variable to plot (X axis label)}

\item{sampleColour}{(str) NULL or vector colour for each sample (same length as \code{apexValue}, \code{widthMin}, \code{widthMax})}

\item{verbose}{(bool) if TRUE message when NA scans are removed}
}
\value{
Grob (ggplot object)
}
\description{
plot for a ROI the value and peakwidth (RT, m/z, ...) of detected peaks across multiple samples (x axis is the value (RT, m/z, ...), y is plotted in the order spectra are passed with the first on top).
}
