% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPantheR_loadAnnotationParamsCSV.R
\name{prepare_basic_target_parameters}
\alias{prepare_basic_target_parameters}
\title{Process target region parameters (simple format) for object
initialisation}
\usage{
prepare_basic_target_parameters(paramTable)
}
\arguments{
\item{paramTable}{(data.frame) Target region parameters}
}
\value{
(list) List of \code{targetFeatTable, uROI, FIR, uROIExist}
}
\description{
Process the simple target region parameters (\code{cpdID, cpdName, mzMin,
mzMax, mz, rtMin, rtMax, rt}) and return input variables for
\code{peakPantheRAnnotation()}
}
