% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{ROI,peakPantheRAnnotation-method}
\alias{ROI,peakPantheRAnnotation-method}
\alias{ROI}
\title{ROI accessor returns targetFeatTable with cpdID, cpdName added}
\usage{
\S4method{ROI}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(data.frame) target feature table with compounds as row and ROI parameters as columns
}
\description{
ROI accessor returns targetFeatTable with cpdID, cpdName added
}
