% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_utils.R
\name{spectra_metadata_colourScheme_UI_helper}
\alias{spectra_metadata_colourScheme_UI_helper}
\title{UI export plot helper - sample colour}
\usage{
spectra_metadata_colourScheme_UI_helper(
    annot,
    splColrColumn = NULL
)
}
\arguments{
\item{annot}{(peakPantheRAnnotation) Annotation object}

\item{splColrColumn}{(str) NULL, None or a spectraMetadata column for
colouring each sample}
}
\value{
(character) Vector of colours
}
\description{
Return a vector of spectra colours based on a metadata column
}
