% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_peakPantheRAnnotation.R
\docType{methods}
\name{uROI,peakPantheRAnnotation-method}
\alias{uROI,peakPantheRAnnotation-method}
\alias{uROI}
\title{uROI accessor returns targetFeatTable with cpdID, cpdName added}
\usage{
\S4method{uROI}{peakPantheRAnnotation}(object)
}
\arguments{
\item{object}{peakPantheRAnnotation}
}
\value{
(data.frame) target feature table with compounds as row and uROI parameters as columns
}
\description{
uROI accessor returns targetFeatTable with cpdID, cpdName added
}
